@@ -20,16 +20,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-builders/custom-webpack": "^7.0.2",
+    "@angular-builders/custom-webpack": "^8.4.1",
     "@angular-builders/dev-server": "^7.3.1",
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
+    "@angular/animations": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "@ionic-native/app-version": "^5.15.1",
     "@ionic-native/core": "^5.15.1",
     "@ionic-native/device": "^5.15.1",
@@ -150,7 +146,7 @@
     "sonarqube-scanner": "^2.5.0",
     "ts-node": "^8.4.1",
     "tslint": "~5.16.0",
-    "typescript": "~3.1.6"
+    "typescript": "~3.4.3"
   },
   "description": "An Ionic project",
   "cordova": {
