@@ -1,6 +1,6 @@
 import { Component, Inject, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { AppGlobalService, AppHeaderService, CommonUtilService, ContentAggregatorHandler, SunbirdQRScanner } from '@app/services';
-import { CourseCardGridTypes, LibraryCardTypes, PillShape, PillsViewType, SelectMode, ButtonPosition, ShowMoreViewType, PillsMultiRow } from '@project-sunbird/common-consumption';
+import { CourseCardGridTypes, LibraryCardTypes, PillShape, PillsViewType, SelectMode, ButtonPosition, ShowMoreViewType, PillsMultiRow } from '@project-sunbird/common-consumption-v8';
 import { NavigationExtras, Router } from '@angular/router';
 import { FrameworkService, FrameworkDetailsRequest, FrameworkCategoryCodesGroup,
   Framework, Profile, ProfileService, ContentAggregatorRequest, ContentSearchCriteria,
@@ -39,7 +39,7 @@ export class UserHomePage implements OnInit, OnDestroy {
   pillsViewType = PillsViewType;
   selectMode = SelectMode;
   pillShape = PillShape;
-  @ViewChild('contentView') contentView: ContentView;
+  @ViewChild('contentView', { static: false }) contentView: ContentView;
   showPreferenceInfo = false;
 
   LibraryCardTypes = LibraryCardTypes
